fix: revert chat-handler.ts refactoring causing timeouts
The modular refactoring of chat-handler.ts introduced in #452 is causing
response timeouts in production. This reverts only the chat handler
changes while keeping other cleanup (md files, Better Auth removal,
shared utilities).

The refactored modules will need more testing before reintroduction.
