+++ apps/web/src/app/api/chat/chat-handler.ts
@@ -1,5 +1,6 @@
 import { convertToCoreMessages, smoothStream, streamText, type UIMessage } from "ai";
 import { createOpenRouter } from "@openrouter/ai-sdk-provider";
+import { createHash } from "crypto";
 import type { Id } from "@server/convex/_generated/dataModel";
 
 import { captureServerEvent } from "@/lib/posthog-server";
@@ -8,96 +9,131 @@ import { ensureConvexUser, streamUpsertMessage } from "@/lib/convex-server";
 import { resolveAllowedOrigins, validateRequestOrigin } from "@/lib/request-origin";
 import { createLogger } from "@/lib/logger";
 import { toConvexUserId, toConvexChatId, toConvexStorageId } from "@/lib/type-converters";
+import type { IRateLimiter } from "@/lib/rate-limit";
 import { hasReasoningCapability } from "@/lib/model-capabilities";
+import { isTextPart, isFilePart } from "@/lib/error-handling";
 import { removeEmDashes } from "@/lib/text-transforms";
 import { EM_DASH_PREVENTION_SYSTEM_PROMPT } from "@/lib/jon-mode-prompts";
 
-// Import from modular files
-import {
-	DEFAULT_RATE_LIMIT,
-	RATE_LIMIT_WINDOW_MS,
-	MAX_TRACKED_RATE_BUCKETS,
-	MAX_MESSAGES_PER_REQUEST,
-	MAX_REQUEST_BODY_SIZE,
-	MAX_ATTACHMENT_SIZE,
-	MAX_MESSAGE_CONTENT_LENGTH,
-	MAX_TOKENS,
-	OPENROUTER_TIMEOUT_MS,
-	OPENROUTER_BASE_URL,
-	STREAM_SMOOTH_DELAY_MS,
-} from "./config";
+/**
+ * Message metadata type with optional createdAt timestamp
+ */
+type MessageMetadata = {
+	createdAt?: string | Date;
+};
 
-import {
-	type AnyUIMessage,
-	type ChatRequestPayload,
-	isReasoningDeltaWithText,
-	extractClientIP,
-	generateClientHash,
-	getErrorStatusCode,
-	validateContentLength,
-	validateBodySize,
-	parseRequestBody,
-	clampUserText,
-	extractMessageText,
-	coerceIsoDate,
-	validateMessageContent,
-} from "./validation";
+type AnyUIMessage = UIMessage<MessageMetadata>;
 
-import {
-	type ChatAttachment,
-	processAttachments,
-	toPersistableAttachments,
-} from "./attachments";
+/**
+ * Type guard to check if a chunk is a reasoning-delta with text
+ */
+function isReasoningDeltaWithText(chunk: { type: string }): chunk is { type: "reasoning-delta"; text: string } {
+	return chunk.type === "reasoning-delta" && "text" in chunk && typeof (chunk as { text?: unknown }).text === "string";
+}
 
-import {
-	type StreamPersistRequest,
-	createStreamManager,
-	createStreamingHeaders,
-} from "./streaming";
-
-// ============================================================================
-// Types
-// ============================================================================
+/**
+ * Helper to extract statusCode from an error object safely
+ */
+function getErrorStatusCode(error: unknown): number | undefined {
+	if (error && typeof error === "object" && "statusCode" in error) {
+		const statusCode = (error as { statusCode: unknown }).statusCode;
+		return typeof statusCode === "number" ? statusCode : undefined;
+	}
+	return undefined;
+}
 
 type RateBucket = {
 	count: number;
 	resetAt: number;
 };
 
-export type ChatHandlerOptions = {
-	streamTextImpl?: typeof streamText;
-	convertToCoreMessagesImpl?: typeof convertToCoreMessages;
-	provider?: ReturnType<typeof createOpenRouter>;
-	model?: string;
-	rateLimit?: { limit: number; windowMs: number };
-	now?: () => number;
-	corsOrigin?: string;
-	persistMessage?: (input: StreamPersistRequest) => Promise<{ ok: boolean }>;
-	resolveModel?: (input: {
-		request: Request;
-		payload: ChatRequestPayload;
-	}) => Promise<{ provider: ReturnType<typeof createOpenRouter>; modelId: string }>;
-};
+// Rate limiting configuration
+/** Default maximum requests per minute if not configured via env */
+const DEFAULT_RATE_LIMIT_PER_MINUTE = 30;
+/** Rate limit window duration in milliseconds (1 minute) */
+const RATE_LIMIT_WINDOW_MS = 60_000;
+/** Default maximum number of rate limit buckets to track in memory */
+const DEFAULT_MAX_TRACKED_BUCKETS = 1_000;
 
-// ============================================================================
-// Helper Functions
-// ============================================================================
+// Message content configuration
+/** Maximum characters allowed in user message parts before truncation */
+const DEFAULT_MAX_USER_CHARS = 8_000;
+
+// Stream buffering configuration
+// PERFORMANCE: Increased intervals to reduce DB write frequency during streaming
+// Previously: 80ms/24chars = ~12 writes/sec, Now: 300ms/80chars = ~3 writes/sec
+/** Default interval for flushing buffered stream chunks to database (milliseconds) */
+const DEFAULT_STREAM_FLUSH_INTERVAL_MS = 300;
+/** Default minimum characters required before flushing a stream chunk */
+const DEFAULT_STREAM_MIN_CHARS_PER_FLUSH = 80;
+/** Default delay between smooth stream word chunks (milliseconds) */
+const DEFAULT_STREAM_SMOOTH_DELAY_MS = 10;
+
+const DEFAULT_RATE_LIMIT = Number(process.env.OPENROUTER_RATE_LIMIT_PER_MIN ?? DEFAULT_RATE_LIMIT_PER_MINUTE);
+const RAW_MAX_TRACKED_RATE_BUCKETS = Number(process.env.OPENROUTER_RATE_LIMIT_TRACKED_BUCKETS ?? DEFAULT_MAX_TRACKED_BUCKETS);
+const MAX_TRACKED_RATE_BUCKETS = Number.isFinite(RAW_MAX_TRACKED_RATE_BUCKETS) && RAW_MAX_TRACKED_RATE_BUCKETS > 0
+	? RAW_MAX_TRACKED_RATE_BUCKETS
+	: null;
+const MAX_USER_PART_CHARS = Number(process.env.OPENROUTER_MAX_USER_CHARS ?? DEFAULT_MAX_USER_CHARS);
+
+// Request size limits for security
+const MAX_MESSAGES_PER_REQUEST = (() => {
+	const val = Number(process.env.MAX_MESSAGES_PER_REQUEST ?? 100);
+	return Number.isFinite(val) && val > 0 ? val : 100;
+})();
+const MAX_REQUEST_BODY_SIZE = (() => {
+	const val = Number(process.env.MAX_REQUEST_BODY_SIZE ?? 10_000_000);
+	return Number.isFinite(val) && val > 0 ? val : 10_000_000;
+})();
+const MAX_ATTACHMENT_SIZE = (() => {
+	const val = Number(process.env.MAX_ATTACHMENT_SIZE ?? 5_000_000);
+	return Number.isFinite(val) && val > 0 ? val : 5_000_000;
+})();
+const MAX_MESSAGE_CONTENT_LENGTH = (() => {
+	const val = Number(process.env.MAX_MESSAGE_CONTENT_LENGTH ?? 50_000);
+	return Number.isFinite(val) && val > 0 ? val : 50_000;
+})();
+
+const STREAM_FLUSH_INTERVAL_RAW = Number(process.env.OPENROUTER_STREAM_FLUSH_INTERVAL_MS ?? DEFAULT_STREAM_FLUSH_INTERVAL_MS);
+const STREAM_FLUSH_INTERVAL_MS =
+	Number.isFinite(STREAM_FLUSH_INTERVAL_RAW) && STREAM_FLUSH_INTERVAL_RAW > 0
+		? STREAM_FLUSH_INTERVAL_RAW
+		: DEFAULT_STREAM_FLUSH_INTERVAL_MS;
+const STREAM_MIN_CHARS_PER_FLUSH_RAW = Number(
+	process.env.OPENROUTER_STREAM_MIN_CHARS_PER_FLUSH ?? DEFAULT_STREAM_MIN_CHARS_PER_FLUSH,
+);
+const STREAM_MIN_CHARS_PER_FLUSH =
+	Number.isFinite(STREAM_MIN_CHARS_PER_FLUSH_RAW) && STREAM_MIN_CHARS_PER_FLUSH_RAW > 0
+		? STREAM_MIN_CHARS_PER_FLUSH_RAW
+		: DEFAULT_STREAM_MIN_CHARS_PER_FLUSH;
+const STREAM_SMOOTH_DELAY_MS = (() => {
+	const raw = process.env.OPENROUTER_STREAM_DELAY_MS;
+	if (raw === undefined || raw === null || raw.trim() === "") return DEFAULT_STREAM_SMOOTH_DELAY_MS;
+	if (raw.trim().toLowerCase() === "null") return null;
+	const parsed = Number(raw);
+	if (!Number.isFinite(parsed)) return DEFAULT_STREAM_SMOOTH_DELAY_MS;
+	return parsed < 0 ? 0 : parsed;
+})();
+const MAX_TOKENS = (() => {
+	const raw = process.env.OPENROUTER_MAX_TOKENS;
+	if (!raw || raw.trim() === "") return 8192;
+	const parsed = Number(raw);
+	if (!Number.isFinite(parsed) || parsed <= 0) return 8192;
+	return Math.min(parsed, 32768); // Cap at 32k to prevent excessive token usage
+})();
+
 
 /**
- * Build CORS headers for response
+ * Convert reasoning config from client to OpenRouter API format
+ *
+ * OpenRouter unified reasoning API supports:
+ * - reasoning.enabled: boolean - Enable reasoning with defaults
+ * - reasoning.effort: "low" | "medium" | "high" - For OpenAI/Grok models
+ * - reasoning.max_tokens: number - For Anthropic/Gemini models
+ * - reasoning.exclude: boolean - Use reasoning but don't include in response
+ *
+ * Based on: https://openrouter.ai/docs/use-cases/reasoning-tokens
  */
-function buildCorsHeaders(request: Request, allowedOrigin?: string | null) {
-	const headers = new Headers();
-	if (allowedOrigin) {
-		headers.set("Access-Control-Allow-Origin", allowedOrigin);
-	}
-	headers.set("Access-Control-Allow-Headers", "Content-Type, Authorization");
-	headers.set("Access-Control-Allow-Methods", "POST,OPTIONS");
-	headers.set("Access-Control-Allow-Credentials", "true");
-	headers.set("Vary", "Origin");
-	return headers;
-}
-
 /**
  * Convert reasoning config from client to OpenRouter API format
  *
@@ -110,6 +146,8 @@ function buildReasoningParam(
 	config: ChatRequestPayload["reasoningConfig"],
 	_modelId: string
 ): Record<string, unknown> | undefined {
+	// When disabled, don't send any reasoning parameter at all
+	// This prevents the model from thinking/reasoning entirely
 	if (!config || !config.enabled) {
 		return undefined;
 	}
@@ -136,9 +174,157 @@ function buildReasoningParam(
 	return reasoning;
 }
 
-// ============================================================================
-// Main Handler
-// ============================================================================
+// PERFORMANCE FIX: Timeout for OpenRouter stream to prevent hanging requests
+const OPENROUTER_TIMEOUT_MS = (() => {
+	const raw = process.env.OPENROUTER_TIMEOUT_MS;
+	if (!raw || raw.trim() === "") return 120000; // Default 2 minutes
+	const parsed = Number(raw);
+	if (!Number.isFinite(parsed) || parsed <= 0) return 120000;
+	return Math.min(parsed, 300000); // Cap at 5 minutes to prevent indefinite hangs
+})();
+
+type StreamPersistRequest = {
+	userId: string;
+	chatId: string;
+	clientMessageId?: string | null;
+	role: "user" | "assistant";
+	content: string;
+	reasoning?: string;
+	thinkingTimeMs?: number;
+	createdAt: string;
+	status: "streaming" | "completed";
+	attachments?: Array<{
+		storageId: Id<"_storage">;
+		filename: string;
+		contentType: string;
+		size: number;
+		uploadedAt: number;
+		url?: string;
+	}>;
+};
+
+type ChatRequestPayload = {
+	modelId?: string;
+	apiKey?: string;
+	chatId?: string;
+	messages?: AnyUIMessage[];
+	assistantMessageId?: string;
+	attachments?: Array<{
+		storageId: Id<"_storage">;
+		filename: string;
+		contentType: string;
+		size: number;
+		url?: string;
+	}>;
+	reasoningConfig?: {
+		enabled: boolean;
+		effort?: "medium" | "high";
+		max_tokens?: number;
+		exclude?: boolean;
+	};
+	jonMode?: boolean;
+};
+
+function clampUserText(message: AnyUIMessage): AnyUIMessage {
+	if (message.role !== "user") return message;
+	let remaining = MAX_USER_PART_CHARS;
+	const parts = message.parts.map((part) => {
+		if (part?.type !== "text") return part;
+		if (remaining <= 0) return { ...part, text: "" };
+		if (part.text.length <= remaining) {
+			remaining -= part.text.length;
+			return part;
+		}
+		const slice = part.text.slice(0, remaining);
+		remaining = 0;
+		return { ...part, text: slice };
+	});
+	return { ...message, parts };
+}
+
+function extractMessageText(message: AnyUIMessage): string {
+	const segments: string[] = [];
+	for (const part of message.parts ?? []) {
+		if (!part) continue;
+		if (isTextPart(part)) {
+			segments.push(part.text);
+			continue;
+		}
+		if (isFilePart(part)) {
+			const name = part.filename && part.filename.length > 0 ? part.filename : "attachment";
+			const media = part.mediaType && part.mediaType.length > 0 ? ` (${part.mediaType})` : "";
+			segments.push(`[Attachment: ${name}${media}]`);
+			continue;
+		}
+	}
+	return segments.join("");
+}
+
+function coerceIsoDate(value: unknown): string {
+	if (typeof value === "string" && value.length > 0) {
+		const date = new Date(value);
+		if (!Number.isNaN(date.valueOf())) return date.toISOString();
+	}
+	if (value instanceof Date && !Number.isNaN(value.valueOf())) {
+		return value.toISOString();
+	}
+	return new Date().toISOString();
+}
+
+function pickClientIp(request: Request): string {
+	const forwarded = request.headers.get("x-forwarded-for");
+	if (forwarded) return forwarded.split(",")[0]!.trim();
+	try {
+		const url = new URL(request.url);
+		return url.hostname;
+	} catch {
+		return "127.0.0.1";
+	}
+}
+
+/**
+ * Anonymize client IP using SHA-256 hash.
+ * 
+ * Security strategy:
+ * - Uses full SHA-256 hash (64 hex chars) for better anonymization
+ * - One-way hash prevents reverse-lookup of original IP
+ * - Allows correlation of requests from same IP without storing PII
+ * - Sufficient for rate limiting and abuse detection
+ */
+function hashClientIp(ip: string): string {
+	try {
+		return createHash("sha256").update(ip).digest("hex");
+	} catch {
+		return "unknown";
+	}
+}
+
+function buildCorsHeaders(request: Request, allowedOrigin?: string | null) {
+	const headers = new Headers();
+	if (allowedOrigin) {
+		headers.set("Access-Control-Allow-Origin", allowedOrigin);
+	}
+	headers.set("Access-Control-Allow-Headers", "Content-Type, Authorization");
+	headers.set("Access-Control-Allow-Methods", "POST,OPTIONS");
+	headers.set("Access-Control-Allow-Credentials", "true");
+	headers.set("Vary", "Origin");
+	return headers;
+}
+
+export type ChatHandlerOptions = {
+	streamTextImpl?: typeof streamText;
+	convertToCoreMessagesImpl?: typeof convertToCoreMessages;
+	provider?: ReturnType<typeof createOpenRouter>;
+	model?: string;
+	rateLimit?: { limit: number; windowMs: number };
+	now?: () => number;
+	corsOrigin?: string;
+	persistMessage?: (input: StreamPersistRequest) => Promise<{ ok: boolean }>;
+	resolveModel?: (input: {
+		request: Request;
+		payload: ChatRequestPayload;
+	}) => Promise<{ provider: ReturnType<typeof createOpenRouter>; modelId: string }>;
+};
 
 export function createChatHandler(options: ChatHandlerOptions = {}) {
 	const logger = createLogger("ChatHandler");
@@ -163,16 +349,9 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				thinkingTimeMs: input.thinkingTimeMs,
 				status: input.status,
 				createdAt: new Date(input.createdAt).getTime(),
-				attachments: input.attachments as Array<{
-					storageId: Id<"_storage">;
-					filename: string;
-					contentType: string;
-					size: number;
-					uploadedAt: number;
-				}> | undefined,
+				attachments: input.attachments,
 			}));
 
-	// Rate limiting state
 	const buckets = new Map<string, RateBucket>();
 	let lastBucketCleanup = 0;
 
@@ -199,7 +378,7 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 
 	function isRateLimited(request: Request): boolean {
 		if (rateLimit.limit <= 0) return false;
-		const ip = extractClientIP(request);
+		const ip = pickClientIp(request);
 		const ts = now();
 		cleanupBuckets(ts);
 		const bucket = buckets.get(ip);
@@ -237,11 +416,10 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 			image: session.image,
 		});
 		const distinctId = session.userId;
-		const clientIp = extractClientIP(request);
-		const ipHash = generateClientHash(clientIp);
+		const clientIp = pickClientIp(request);
+		const ipHash = hashClientIp(clientIp);
 		const requestOriginValue = originResult.origin ?? request.headers.get("origin") ?? allowOrigin ?? null;
 		const rateLimitBucketLabel = `${rateLimit.limit}/${bucketWindowMs}`;
-
 		if (isRateLimited(request)) {
 			const headers = buildCorsHeaders(request, allowOrigin);
 			headers.set("Retry-After", Math.ceil(bucketWindowMs / 1000).toString());
@@ -259,10 +437,16 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 		}
 
 		// Validate request body size BEFORE loading into memory
-		const contentLengthResult = validateContentLength(request);
-		if (!contentLengthResult.isValid) {
-			const headers = buildCorsHeaders(request, allowOrigin);
-			return new Response(contentLengthResult.error, { status: contentLengthResult.statusCode ?? 413, headers });
+		const contentLength = request.headers.get("content-length");
+		if (contentLength) {
+			const declaredSize = Number.parseInt(contentLength, 10);
+			if (!Number.isNaN(declaredSize) && declaredSize > MAX_REQUEST_BODY_SIZE) {
+				const headers = buildCorsHeaders(request, allowOrigin);
+				return new Response(
+					`Request body too large: ${declaredSize} bytes (max: ${MAX_REQUEST_BODY_SIZE} bytes)`,
+					{ status: 413, headers }
+				);
+			}
 		}
 
 		let payload: ChatRequestPayload;
@@ -270,20 +454,17 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 		try {
 			rawBody = await request.text();
 
-			// Double-check actual body size after loading
+			// Double-check actual body size after loading (in case Content-Length was missing or incorrect)
 			const bodySize = new TextEncoder().encode(rawBody).length;
-			const bodySizeResult = validateBodySize(bodySize);
-			if (!bodySizeResult.isValid) {
+			if (bodySize > MAX_REQUEST_BODY_SIZE) {
 				const headers = buildCorsHeaders(request, allowOrigin);
-				return new Response(bodySizeResult.error, { status: bodySizeResult.statusCode ?? 413, headers });
+				return new Response(
+					`Request body too large: ${bodySize} bytes (max: ${MAX_REQUEST_BODY_SIZE} bytes)`,
+					{ status: 413, headers }
+				);
 			}
 
-			const parseResult = parseRequestBody(rawBody);
-			if (!parseResult.payload) {
-				const headers = buildCorsHeaders(request, allowOrigin);
-				return new Response(parseResult.error ?? "Invalid request", { status: parseResult.statusCode ?? 400, headers });
-			}
-			payload = parseResult.payload;
+			payload = JSON.parse(rawBody);
 		} catch (error: unknown) {
 			const headers = buildCorsHeaders(request, allowOrigin);
 			if (error instanceof SyntaxError) {
@@ -292,7 +473,6 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 			throw error;
 		}
 
-		// Resolve model configuration
 		const modelIdFromPayload = typeof payload?.modelId === "string" && payload.modelId.trim().length > 0
 			? payload.modelId.trim()
 			: null;
@@ -315,7 +495,7 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				}
 				const provider = createOpenRouter({
 					apiKey: apiKeyFromPayload,
-					baseURL: OPENROUTER_BASE_URL,
+					baseURL: process.env.OPENROUTER_BASE_URL || "https://openrouter.ai/api/v1",
 					compatibility: "strict",
 				});
 				config = { provider, modelId: modelIdFromPayload };
@@ -328,6 +508,7 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 			let status = 500;
 			let responseMessage = "Server configuration error";
 
+			// Check for specific known errors
 			if (messageText.includes("Missing apiKey")) {
 				status = 400;
 				responseMessage = "Missing apiKey";
@@ -335,26 +516,26 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				status = 400;
 				responseMessage = "Missing modelId";
 			} else if (!isProduction) {
+				// In development, include detailed error for debugging
 				responseMessage = messageText || "Server configuration error";
 			}
 
 			return new Response(responseMessage, { status, headers });
 		}
 
-		// Validate chatId
 		const chatId = typeof payload?.chatId === "string" && payload.chatId.trim().length > 0 ? payload.chatId.trim() : null;
 		if (!chatId) {
 			const headers = buildCorsHeaders(request, allowOrigin);
 			return new Response("Missing chatId", { status: 400, headers });
 		}
 
-		// Validate messages
 		const rawMessages: AnyUIMessage[] = Array.isArray(payload?.messages) ? (payload.messages as AnyUIMessage[]) : [];
 		if (rawMessages.length === 0) {
 			const headers = buildCorsHeaders(request, allowOrigin);
 			return new Response("Missing chat messages", { status: 400, headers });
 		}
 
+		// Validate messages array length
 		if (rawMessages.length > MAX_MESSAGES_PER_REQUEST) {
 			const headers = buildCorsHeaders(request, allowOrigin);
 			return new Response(
@@ -363,20 +544,74 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 			);
 		}
 
-		// Validate message content lengths and attachment sizes
-		const contentValidation = validateMessageContent(rawMessages);
-		if (!contentValidation.isValid) {
-			const headers = buildCorsHeaders(request, allowOrigin);
-			return new Response(contentValidation.error, { status: contentValidation.statusCode ?? 413, headers });
+		// Validate message content length and attachment sizes
+		for (let i = 0; i < rawMessages.length; i++) {
+			const message = rawMessages[i];
+			if (!message) continue;
+
+			// Check each part of the message
+			for (const part of message.parts ?? []) {
+				if (!part) continue;
+
+				// Validate text content length
+				if (isTextPart(part)) {
+					const textLength = part.text.length;
+					if (textLength > MAX_MESSAGE_CONTENT_LENGTH) {
+						const headers = buildCorsHeaders(request, allowOrigin);
+						return new Response(
+							`Message content too long: ${textLength} characters (max: ${MAX_MESSAGE_CONTENT_LENGTH})`,
+							{ status: 413, headers }
+						);
+					}
+				}
+
+				// Validate attachment size - check combined size to prevent bypass
+				if (isFilePart(part)) {
+					let totalAttachmentSize = 0;
+
+					// Add data URL size if present
+					if (typeof part.data === "string") {
+						totalAttachmentSize += new TextEncoder().encode(part.data).length;
+					}
+
+					// Add url content size if it's a data URL
+					if (typeof part.url === "string" && part.url.startsWith("data:")) {
+						totalAttachmentSize += new TextEncoder().encode(part.url).length;
+					}
+
+					// Check combined size to prevent bypass by splitting content
+					if (totalAttachmentSize > MAX_ATTACHMENT_SIZE) {
+						const headers = buildCorsHeaders(request, allowOrigin);
+						return new Response(
+							`Attachment too large: ${totalAttachmentSize} bytes (max: ${MAX_ATTACHMENT_SIZE} bytes)`,
+							{ status: 413, headers }
+						);
+					}
+				}
+			}
 		}
 
 		// Convert Convex attachments to AI SDK file parts
 		if (payload.attachments && payload.attachments.length > 0) {
 			const { getFileUrl } = await import("@/lib/convex-server");
 
-			const fileParts = await processAttachments(
-				payload.attachments as ChatAttachment[],
-				(storageId) => getFileUrl(toConvexStorageId(storageId as unknown as Id<"_storage">), convexUserId)
+			// Fetch URLs for all attachments
+			const fileParts = await Promise.all(
+				payload.attachments.map(async (attachment) => {
+					const url = await getFileUrl(
+						toConvexStorageId(attachment.storageId),
+						convexUserId,
+					);
+					if (!url) {
+						throw new Error(`Failed to get URL for file: ${attachment.filename}`);
+					}
+					return {
+						type: "file" as const,
+						mediaType: attachment.contentType,
+						url,
+						filename: attachment.filename,
+					};
+				}),
 			);
 
 			// Add file parts to the last user message
@@ -404,6 +639,7 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 		const userCreatedAtIso = coerceIsoDate(rawUserMessage.metadata?.createdAt);
 		const userContent = extractMessageText(safeUserMessage);
 		if (safeUserMessage.id !== userMessageId) {
+			// UIMessage.id is mutable, safe to assign directly
 			(safeUserMessage as { id: string }).id = userMessageId;
 		}
 
@@ -412,7 +648,8 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 			: (crypto.randomUUID?.() ?? `assistant-${Date.now()}`);
 		const assistantCreatedAtIso = new Date().toISOString();
 
-		// Persist user message and bootstrap assistant message in PARALLEL
+		// PERFORMANCE: Persist user message and bootstrap assistant message in PARALLEL
+		// This cuts pre-stream latency in half by avoiding sequential DB round-trips
 		try {
 			const [userResult, assistantBootstrap] = await Promise.all([
 				persistMessageImpl({
@@ -423,10 +660,13 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 					content: userContent,
 					createdAt: userCreatedAtIso,
 					status: "completed",
-					attachments: toPersistableAttachments(
-						payload.attachments as ChatAttachment[] | undefined,
-						toConvexStorageId
-					),
+					attachments: payload.attachments?.map(a => ({
+						storageId: toConvexStorageId(a.storageId),
+						filename: a.filename,
+						contentType: a.contentType,
+						size: a.size,
+						uploadedAt: Date.now(),
+					})),
 				}),
 				persistMessageImpl({
 					userId: convexUserId,
@@ -451,20 +691,141 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 			return new Response("Failed to persist messages", { status: 502, headers });
 		}
 
-		// Create stream manager for buffered persistence
-		const streamManager = createStreamManager({
-			userId: convexUserId,
-			chatId,
-			assistantMessageId,
-			assistantCreatedAtIso,
-			persistMessage: persistMessageImpl,
-		});
-
+		let assistantText = "";
+		let assistantReasoning = "";
+		let lastPersistedLength = 0;
+		let lastPersistedReasoningLength = 0;
+		let flushTimeout: ReturnType<typeof setTimeout> | null = null;
+		let pendingFlush: Promise<void> | null = null;
+		let pendingResolve: (() => void) | null = null;
+		let finalized = false;
+		let persistenceError: Error | null = null;
 		const startedAt = Date.now();
+		let streamStatus: "completed" | "aborted" | "error" = "completed";
+		let reasoningStartTime: number | null = null;
+		let reasoningEndTime: number | null = null;
 
-		// Create AbortController for timeout protection only
+		const persistAssistant = async (status: "streaming" | "completed", force = false) => {
+			const pendingLength = assistantText.length;
+			const pendingReasoningLength = assistantReasoning.length;
+			const delta = pendingLength - lastPersistedLength;
+			const reasoningDelta = pendingReasoningLength - lastPersistedReasoningLength;
+			if (!force && status === "streaming") {
+				if (delta <= 0 && reasoningDelta <= 0) return;
+				if (delta < STREAM_MIN_CHARS_PER_FLUSH && reasoningDelta < STREAM_MIN_CHARS_PER_FLUSH) return;
+			}
+			if (delta <= 0 && reasoningDelta <= 0 && !force) {
+				return;
+			}
+			lastPersistedLength = pendingLength;
+			lastPersistedReasoningLength = pendingReasoningLength;
+
+			// CRITICAL: Only send reasoning if it's non-empty
+			const reasoningToSend = assistantReasoning.length > 0 ? assistantReasoning : undefined;
+
+			// Calculate thinking time duration
+			const thinkingTimeMs = reasoningStartTime && reasoningEndTime
+				? reasoningEndTime - reasoningStartTime
+				: undefined;
+
+			if (status === "completed") {
+				logger.debug("Final save - Reasoning status", {
+					textLength: assistantText.length,
+					reasoningLength: reasoningToSend ? reasoningToSend.length : 0,
+					hasReasoning: !!reasoningToSend,
+					thinkingTimeMs,
+					thinkingTimeSec: thinkingTimeMs ? (thinkingTimeMs / 1000).toFixed(1) : 0,
+					reasoningPreview: reasoningToSend ? reasoningToSend.slice(0, 100) : "NONE"
+				});
+			}
+
+			const response = await persistMessageImpl({
+				userId: convexUserId,
+				chatId,
+				clientMessageId: assistantMessageId,
+				role: "assistant",
+				content: assistantText,
+				reasoning: reasoningToSend,
+				thinkingTimeMs,
+				createdAt: assistantCreatedAtIso,
+				status,
+			});
+			if (!response.ok) {
+				throw new Error("assistant streamUpsert rejected");
+			}
+		};
+
+		const scheduleStreamFlush = () => {
+			if (flushTimeout) return pendingFlush ?? Promise.resolve();
+			const promise = new Promise<void>((resolve) => {
+				pendingResolve = resolve;
+			});
+			pendingFlush = promise;
+			flushTimeout = setTimeout(async () => {
+				flushTimeout = null;
+				try {
+					await persistAssistant("streaming");
+				} catch (error: unknown) {
+					logger.error("Failed to persist assistant chunk", error);
+					if (!persistenceError) {
+						persistenceError = error instanceof Error ? error : new Error("Failed to persist assistant chunk");
+					}
+				} finally {
+					pendingResolve?.();
+					pendingResolve = null;
+					pendingFlush = null;
+				}
+			}, STREAM_FLUSH_INTERVAL_MS);
+			return promise;
+		};
+
+		const finalize = async () => {
+			if (finalized) return;
+			finalized = true;
+			if (flushTimeout) {
+				clearTimeout(flushTimeout);
+				flushTimeout = null;
+			}
+			const pending = pendingFlush;
+			if (pendingResolve) {
+				pendingResolve();
+				pendingResolve = null;
+			}
+			pendingFlush = null;
+			if (pending) {
+			try {
+				await pending;
+			} catch (error: unknown) {
+				logger.error("Failed to await pending flush during finalize", error);
+			}
+			}
+			try {
+				await persistAssistant("completed", true);
+			} catch (error: unknown) {
+				logger.error("Failed to persist assistant completion", error);
+				if (!persistenceError) {
+					persistenceError = error instanceof Error ? error : new Error("Failed to persist assistant completion");
+				}
+			}
+			if (persistenceError) {
+				throw persistenceError;
+			}
+		};
+
+		// STREAM PERSISTENCE: Create AbortController only for timeout protection.
+		// We intentionally DO NOT forward request.signal (browser connection close) to the abort controller.
+		// This ensures the stream continues running server-side even when the user:
+		// - Closes the tab
+		// - Reloads the page
+		// - Loses internet connection
+		// - Switches tabs/apps
+		// The stream will only abort on timeout or explicit stop (handled separately).
+		// Combined with consumeStream(), this guarantees the message is always saved.
 		const abortController = new AbortController();
 		let timeoutId: ReturnType<typeof setTimeout> | null = null;
+
+		// NOTE: We no longer forward request.signal abort to allow stream persistence.
+		// The stream continues running via consumeStream() even if the client disconnects.
 
 		try {
 			timeoutId = setTimeout(() => {
@@ -473,9 +834,14 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 			}, OPENROUTER_TIMEOUT_MS);
 
 			const hasReasoning = hasReasoningCapability(config.modelId);
+
+			// Build reasoning parameter from client configuration using OpenRouter unified API
 			const reasoning = buildReasoningParam(payload.reasoningConfig, config.modelId);
+
+			// Jon Mode: Extract em-dash prevention setting
 			const jonMode = payload.jonMode ?? false;
 
+			// DEBUG: Log reasoning config to understand what's being sent
 			logger.debug("Reasoning configuration", {
 				modelId: config.modelId,
 				hasReasoningCapability: hasReasoning,
@@ -484,6 +850,8 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				willPassExtraBody: !!reasoning,
 			});
 
+			// CRITICAL: Pass reasoning via extraBody when creating the model per OpenRouter AI SDK docs
+			// When reasoning is undefined (disabled), NO extraBody is passed at all
 			const model = config.provider.chat(config.modelId, {
 				...(reasoning && {
 					extraBody: {
@@ -508,56 +876,78 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				messages: convertToCoreMessagesImpl(messagesForModel),
 				maxOutputTokens: MAX_TOKENS,
 				abortSignal: abortController.signal,
+				// CRITICAL FIX: Don't apply smoothStream to reasoning models - it blocks reasoning chunks!
 				experimental_transform: hasReasoning && reasoning ? undefined : smoothStream({
 					delayInMs: STREAM_SMOOTH_DELAY_MS,
 					chunking: "word",
 				}),
 				onChunk: async ({ chunk }) => {
 					if (chunk.type === "text-delta" && chunk.text.length > 0) {
+						// Jon Mode: Post-processing failsafe - remove em-dashes from chunks
 						const textToAdd = jonMode ? removeEmDashes(chunk.text) : chunk.text;
-						streamManager.appendText(textToAdd);
-						streamManager.scheduleFlush();
+						assistantText += textToAdd;
+						scheduleStreamFlush();
 					} else if (isReasoningDeltaWithText(chunk)) {
+						// Start timer on FIRST reasoning chunk if not started
+						if (!reasoningStartTime) {
+							reasoningStartTime = Date.now();
+							logger.debug("Reasoning started", { startTime: reasoningStartTime });
+						}
+
 						if (chunk.text.length > 0) {
-							streamManager.appendReasoning(chunk.text);
-							streamManager.scheduleFlush();
+							assistantReasoning += chunk.text;
+							// Update end time on EVERY chunk (captures last one)
+							reasoningEndTime = Date.now();
+							scheduleStreamFlush();
 						}
 					}
 				},
 				onFinish: async (_event) => {
 					if (timeoutId) clearTimeout(timeoutId);
+					streamStatus = "completed";
+
+					// CRITICAL FIX: If we have reasoning but no end time, set it now
+					if (assistantReasoning.length > 0 && !reasoningEndTime) {
+						reasoningEndTime = Date.now();
+					}
+
+					const duration = reasoningStartTime && reasoningEndTime
+						? reasoningEndTime - reasoningStartTime
+						: 0;
 					logger.debug("Stream finished", {
-						textLength: streamManager.state.fullText.length,
-						reasoningLength: streamManager.state.reasoningText.length,
-						hasReasoning: streamManager.state.reasoningText.length > 0,
+						textLength: assistantText.length,
+						reasoningLength: assistantReasoning.length,
+						hasReasoning: assistantReasoning.length > 0,
+						thinkingTimeMs: duration,
 					});
-					await streamManager.finalize();
+					await finalize();
 				},
 				onAbort: async () => {
 					if (timeoutId) clearTimeout(timeoutId);
-					streamManager.markAborted();
-					await streamManager.finalize();
+					streamStatus = "aborted";
+					await finalize();
 				},
 				onError: async ({ error }) => {
 					if (timeoutId) clearTimeout(timeoutId);
 					logger.error("Chat stream error", error);
-					streamManager.markError();
-					await streamManager.finalize();
+					streamStatus = "error";
+					await finalize();
 				},
 			});
-
-			// Consume the stream to ensure it runs to completion
+			// CRITICAL: Consume the stream to ensure it runs to completion and triggers onFinish
+			// even when the client aborts the connection (tab close, reload, disconnect).
+			// This removes backpressure from the LLM provider and guarantees the message is saved.
+			// DO NOT await - fire and forget so the response can be returned immediately.
 			result.consumeStream();
 
 			const duration = Date.now() - startedAt;
-			const streamStatus = streamManager.state.status;
 			const openrouterStatus = streamStatus === "completed" ? "ok" : streamStatus;
 			captureServerEvent("chat_message_stream", distinctId, {
 				chat_id: chatId,
 				model_id: config.modelId,
 				user_message_id: userMessageId,
 				assistant_message_id: assistantMessageId,
-				characters: streamManager.state.fullText.length,
+				characters: assistantText.length,
 				duration_ms: duration,
 				status: streamStatus,
 				openrouter_status: openrouterStatus,
@@ -571,11 +961,16 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				generateMessageId: () => assistantMessageId,
 				sendReasoning: hasReasoning,
 			});
+			const headers = new Headers(aiResponse.headers);
+			headers.set("Cache-Control", "no-store, max-age=0");
+			headers.set("X-Accel-Buffering", "no");
+			headers.set("Connection", "keep-alive");
 
 			const corsHeaders = buildCorsHeaders(request, allowOrigin);
-			const headers = createStreamingHeaders(aiResponse.headers, corsHeaders);
+			corsHeaders.forEach((value, key) => {
+				headers.set(key, value);
+			});
 
-			const persistenceError = streamManager.getPersistenceError();
 			if (persistenceError) {
 				throw persistenceError;
 			}
@@ -585,22 +980,27 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				headers,
 			});
 		} catch (error: unknown) {
+			// Ensure timeout is cleared even if error occurs before callbacks
 			if (timeoutId) clearTimeout(timeoutId);
 			logger.error("Chat handler error", error, { chatId, modelId: config.modelId });
-
 			try {
-				await streamManager.finalize();
+				await finalize();
 			} catch (finalizeError: unknown) {
 				logger.error("Chat finalize error", finalizeError);
+				if (!persistenceError) {
+					persistenceError =
+						finalizeError instanceof Error
+							? finalizeError
+							: new Error(String(finalizeError));
+				}
 			}
-
 			const duration = Date.now() - startedAt;
 			captureServerEvent("chat_message_stream", distinctId, {
 				chat_id: chatId,
 				model_id: config.modelId,
 				user_message_id: userMessageId,
 				assistant_message_id: assistantMessageId,
-				characters: streamManager.state.fullText.length,
+				characters: assistantText.length,
 				duration_ms: duration,
 				status: "error",
 				openrouter_status: "error",
@@ -609,15 +1009,16 @@ export function createChatHandler(options: ChatHandlerOptions = {}) {
 				ip_hash: ipHash,
 				rate_limit_bucket: rateLimitBucketLabel,
 			});
-
 			const headers = buildCorsHeaders(request, allowOrigin);
 			const upstreamStatus = getErrorStatusCode(error);
 			const isProduction = process.env.NODE_ENV === "production";
 
+			// Handle specific error cases with generic messages in production
 			if (upstreamStatus === 401) {
 				return new Response("OpenRouter API key invalid", { status: 401, headers });
 			}
 
+			// In production, return generic error message; in development, include details
 			let errorMessage = "Upstream error";
 			if (!isProduction && error instanceof Error) {
 				errorMessage = `Upstream error: ${error.message}`;
@@ -641,3 +1042,4 @@ export function createOptionsHandler(options: { corsOrigin?: string } = {}) {
 		return new Response(null, { status: 204, headers });
 	};
 }
+
